+++ lib/general.sh
@@ -361,3 +361,9 @@ fi
 shellspec_match() {
   [ "${2:-}" ] && eval "case \${1:-} in ($2) true ;; (*) false ;; esac &&:"
 }
+
+shellspec_join() {
+  IFS=" $IFS"
+  eval "shift; $1=\${*:-}"
+  IFS=${IFS#?}
+}